Add login unit test
